Add possibility to compare results
The argument `--compare` allows the user to compare the instantaneous
lift and drag coefficients with other OpenFoam simulations.
The flag `--compare` should be followed by the path of all other OpenFoam
simulations. There is a limit a 4 other OpenFoam cases right now. The limit
can be increased by adding colors to the `colors` list in the function
`plot_coefficients`. When displaying the figure, the legend can appear outside
the figure; this is a known-bug from matplotlib. However, when saving
the figure, the legend is displayed entirely.