Rewrite way to add boxes and physical surfaces
Writing a box has been pushed into a function.
Now, the script deals with levels of refinement, just like snappyHexMesh.
Instead of giving the number of points on the body, or the characteristic
length, the user gives the level of refinement. From one level to the next one,
the chracteristic length is divided by 2. If you, want to change the buffer
layer size (number of cells of same characteristic length separating two
consecutive levels), modify the variable `n_buffer` into the script
(currently: `n_buffer = 10`).
Also, the physical surfaces are now directly written in the script generating
the .geo file. Therefore, the script `write_physical_surfaces.py` is useless
to generate a GMSH mesh.