Add script to re-discretize a geometry
Given a coordinates file and a number of points (or a characteristic length),
the script uniformely re-discretizes the geometry and saves the new coordinates
into a file.
Because I copy a script from pyIBM, it also allows the user to rotate, scale
and translate the geometry.
See flag `--help` for other command-line arguments.