Change the way to compute mean, extrema and St
First, the mean coefficients are computed within a given time-interval.
Then, the extrema are calculated over the last period of oscillation.
Finally, the Strouhal number is evaluated over the last period.

Why did I change?
Previously, the extrema were calculated over the same time-interval used to
calculate the mean coefficients. In the case where the pseudo-steady state
is not reached yet, this method does not give an enough accurate estimation
of the deviations. The same applies to the Strouhal number.

To be note: the old methods are still in the code.