Add script to analyze force coefficients
Allows use to display the intantaneous force coefficients, and
its mean values, its Strouhal number and its extrema over the time.
This is done for the lift and drag coefficients of a given OpenFoam simulation.

See flag `--help` to have a list of command-line options.

Further work: the script should log the results in a file.